Refactor wrapper render package
