+++ minecraft/1.7/src/main/java/nova/core/wrapper/mc/forge/v17/wrapper/render/RenderUtility.java
@@ -18,7 +18,7 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package nova.core.wrapper.mc.forge.v17.render;
+package nova.core.wrapper.mc.forge.v17.wrapper.render;
 
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
