+++ minecraft/1.8/src/main/java/nova/core/wrapper/mc/forge/v18/wrapper/entity/forward/FWEntityFX.java
@@ -34,7 +34,7 @@ import nova.core.component.transform.EntityTransform;
 import nova.core.entity.Entity;
 import nova.core.entity.EntityFactory;
 import nova.core.util.shape.Cuboid;
-import nova.core.wrapper.mc.forge.v18.render.RenderUtility;
+import nova.core.wrapper.mc.forge.v18.wrapper.render.RenderUtility;
 import nova.core.wrapper.mc.forge.v18.util.WrapperEvent;
 import nova.internal.core.Game;
