+++ minecraft/1.8/src/main/java/nova/core/wrapper/mc/forge/v18/wrapper/render/backward/BWModel.java
@@ -18,7 +18,7 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package nova.core.wrapper.mc.forge.v18.wrapper.render;
+package nova.core.wrapper.mc.forge.v18.wrapper.render.backward;
 
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.WorldRenderer;
@@ -29,7 +29,7 @@ import nova.core.render.model.CustomModel;
 import nova.core.render.model.MeshModel;
 import nova.core.render.texture.EntityTexture;
 import nova.core.render.texture.Texture;
-import nova.core.wrapper.mc.forge.v18.render.RenderUtility;
+import nova.core.wrapper.mc.forge.v18.wrapper.render.RenderUtility;
 import nova.core.wrapper.mc.forge.v18.wrapper.assets.AssetConverter;
 
 import java.util.Optional;
@@ -53,8 +53,8 @@ public class BWModel extends MeshModel {
 		render(Optional.empty(), Optional.empty());
 	}
 
-	public void render(Optional<RenderManager> entityRenderManager) {
-		render(Optional.empty(), entityRenderManager);
+	public void render(RenderManager entityRenderManager) {
+		render(Optional.empty(), Optional.of(entityRenderManager));
 	}
 
 	public void render(Optional<IBlockAccess> access, Optional<RenderManager> entityRenderManager) {
